+++ python/cudf/cudf/tests/test_dask.py
@@ -1,17 +1,19 @@
-# Copyright (c) 2019, NVIDIA CORPORATION.
+# Copyright (c) 2019-2024, NVIDIA CORPORATION.
+
+from datetime import datetime
 
 import pytest
 
 import cudf
 
-is_dataframe_like = pytest.importorskip(
-    "dask.dataframe.utils"
-).is_dataframe_like
-is_index_like = pytest.importorskip("dask.dataframe.utils").is_index_like
-is_series_like = pytest.importorskip("dask.dataframe.utils").is_series_like
-
 
 def test_is_dataframe_like():
+    is_dataframe_like = pytest.importorskip(
+        "dask.dataframe.utils"
+    ).is_dataframe_like
+    is_index_like = pytest.importorskip("dask.dataframe.utils").is_index_like
+    is_series_like = pytest.importorskip("dask.dataframe.utils").is_series_like
+
     df = cudf.DataFrame({"x": [1, 2, 3]})
     assert is_dataframe_like(df)
     assert is_series_like(df.x)
@@ -19,3 +21,57 @@ def test_is_dataframe_like():
     assert not is_dataframe_like(df.x)
     assert not is_series_like(df)
     assert not is_index_like(df)
+
+
+@pytest.mark.parametrize("index", [None, [1, 2] * 5])
+def test_deterministic_tokenize(index):
+    tokenize = pytest.importorskip("dask.base").tokenize
+
+    # Checks that `dask.base.normalize_token` correctly
+    # dispatches to the logic defined in `backends.py`
+    # (making `tokenize(<cudf-data>)` deterministic).
+    df = cudf.DataFrame(
+        {"A": range(10), "B": ["dog", "cat"] * 5, "C": range(10, 0, -1)},
+        index=index,
+    )
+
+    # Matching data should produce the same token
+    assert tokenize(df) == tokenize(df)
+    assert tokenize(df.A) == tokenize(df.A)
+    assert tokenize(df.index) == tokenize(df.index)
+    assert tokenize(df) == tokenize(df.copy(deep=True))
+    assert tokenize(df.A) == tokenize(df.A.copy(deep=True))
+    assert tokenize(df.index) == tokenize(df.index.copy(deep=True))
+
+    # Modifying a column element should change the token
+    original_token = tokenize(df)
+    original_token_a = tokenize(df.A)
+    df.A.iloc[2] = 10
+    assert original_token != tokenize(df)
+    assert original_token_a != tokenize(df.A)
+
+    # Modifying an index element should change the token
+    original_token = tokenize(df)
+    original_token_index = tokenize(df.index)
+    new_index = df.index.values
+    new_index[2] = 10
+    df.index = new_index
+    assert original_token != tokenize(df)
+    assert original_token_index != tokenize(df.index)
+
+    # Check MultiIndex case
+    df2 = df.set_index(["B", "C"], drop=False)
+    assert tokenize(df) != tokenize(df2)
+    assert tokenize(df2) == tokenize(df2)
+
+
+def test_deterministic_tokenize_multiindex():
+    tokenize = pytest.importorskip("dask.base").tokenize
+
+    dt = datetime.strptime("1995-03-15", "%Y-%m-%d")
+    index = cudf.MultiIndex(
+        levels=[[1, 2], [dt]],
+        codes=[[0, 1], [0, 0]],
+    )
+    df = cudf.DataFrame(index=index)
+    assert tokenize(df) == tokenize(df)
